Add vote and commentary system

+++ database/migrations/2018_03_26_092412_create_medias_table.php
@@ -1,33 +1,27 @@
 <?php
+
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateMediasTable extends Migration
-{
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
-    {
-        Schema::create('medias', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('file' , 255);
-            $table->integer('template_id')->unsigned();
-            $table->enum('type' , ['image', 'youtube']);
-            $table->foreign('template_id')->references('id')->on('templates')->onDelete('cascade');
-            $table->timestamps();
-        });
+class CreateMediasTable extends Migration {
+
+    public function up() {
+
+        if (!Schema::hasTable('medias')) {
+
+            Schema::create('medias', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('file', 255);
+                $table->integer('template_id')->unsigned();
+                $table->enum('type', ['image', 'youtube']);
+                $table->foreign('template_id')->references('id')->on('templates')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
+
+    public function down() {
         Schema::dropIfExists('medias');
     }
 }
