+++ database/migrations/2018_03_28_124819_create_order_template.php
@@ -8,19 +8,25 @@ class CreateOrderTemplate extends Migration {
 
     public function up() {
 
-        Schema::create('order_template', function (Blueprint $table) {
-            $table->increments('id');
+        if (!Schema::hasTable('order_template')) {
 
-            $table->integer('order_id')->unsigned()->index();
-            $table->integer('template_id')->unsigned()->index();
-            $table->integer('quantity')->unsigned();
+            Schema::create('order_template', function (Blueprint $table) {
+                $table->increments('id');
 
-            $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
-            $table->foreign('template_id')->references('id')->on('templates')->onDelete('cascade');
-        });
+                $table->integer('order_id')->unsigned()->index();
+                $table->integer('template_id')->unsigned()->index();
+                $table->integer('quantity')->unsigned();
+
+                $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
+                $table->foreign('template_id')->references('id')->on('templates')->onDelete('cascade');
+
+            });
+        }
     }
+
+
     public function down() {
-        
+
         Schema::dropIfExists('order_template');
     }
 }
